refactor: Move logic to object and change directory
