+++ src/main/java/com/innim/okkycopy/domain/member/dto/request/SignupRequest.java
@@ -1,5 +1,6 @@
 package com.innim.okkycopy.domain.member.dto.request;
 
+import com.innim.okkycopy.global.error.ErrorCode;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.Pattern;
 import jakarta.validation.constraints.Size;
@@ -30,5 +31,22 @@ public class SignupRequest {
 
     public void encodePassword(PasswordEncoder passwordEncoder) {
         password = passwordEncoder.encode(password);
+    }
+
+    public static ErrorCode retreiveErrorCode(String field) {
+        switch(field) {
+            case("id"):
+                return ErrorCode._400_INVALID_ID;
+            case("password"):
+                return ErrorCode._400_INVALID_PW;
+            case("email"):
+                return ErrorCode._400_INVALID_EMAIL;
+            case("name"):
+                return ErrorCode._400_INVALID_NAME;
+            case("nickname"):
+                return ErrorCode._400_INVALID_NICKNAME;
+            default:
+                return ErrorCode._400_INVALID_UNEXPECTED;
+        }
     }
 }
