+++ src/main/java/com/innim/okkycopy/domain/member/entity/Member.java
@@ -1,7 +1,7 @@
 package com.innim.okkycopy.domain.member.entity;
 
 import com.innim.okkycopy.domain.member.dto.request.SignupRequest;
-import com.innim.okkycopy.global.enums.Role;
+import com.innim.okkycopy.global.auth.enums.Role;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
